extract functions into separate modules
