+++ packages/babel-plugin/package.json
@@ -13,8 +13,9 @@
     "jest": "^23.4.0"
   },
   "dependencies": {
-    "babel-generator": "^6.26.1",
     "@babel/helper-plugin-utils": "^7.0.0-beta.53",
+    "babel-generator": "^6.26.1",
+    "babel-types": "^6.26.0",
     "jss": "^9.8.7",
     "jss-preset-default": "^4.5.0",
     "lodash": "^4.17.10"
