+++ packages/babel-plugin/src/isFromImport.js
@@ -1,0 +1,4 @@
+import * as t from 'babel-types'
+
+export default node =>
+  t.isImportDeclaration(node) || t.isImportSpecifier(node) || t.isImportDefaultSpecifier(node)
